requirements.txt
newsapi-python==0.2.6
Flask==1.1.1
gunicorn==20.0.4
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy>=1.9.2
scipy>=0.15.1
scikit-learn>=0.18
pandas>=0.19
certifi==2019.3.9
chardet==3.0.4
Click==7.0
idna==2.8
itsdangerous==1.1.0
PyJWT==1.7.1
PySocks==1.6.8
pytz==2018.9
requests==2.21.0
six==1.12.0
twilio==6.25.2
urllib3==1.24.1